Move condition/trigger test helpers to test.components.common
