+++ tests/components/alarm_control_panel/test_condition.py
@@ -11,7 +11,7 @@ from homeassistant.components.alarm_control_panel import (
 from homeassistant.const import ATTR_SUPPORTED_FEATURES
 from homeassistant.core import HomeAssistant
 
-from tests.components import (
+from tests.components.common import (
     ConditionStateDescription,
     assert_condition_gated_by_labs_flag,
     create_target_condition,
