+++ tests/components/device_tracker/test_trigger.py
@@ -7,7 +7,7 @@ import pytest
 from homeassistant.const import CONF_ENTITY_ID, STATE_HOME, STATE_NOT_HOME
 from homeassistant.core import HomeAssistant, ServiceCall
 
-from tests.components import (
+from tests.components.common import (
     TriggerStateDescription,
     arm_trigger,
     assert_trigger_behavior_any,
