+++ tests/components/humidifier/test_trigger.py
@@ -8,7 +8,7 @@ from homeassistant.components.humidifier.const import ATTR_ACTION, HumidifierAct
 from homeassistant.const import CONF_ENTITY_ID, STATE_OFF, STATE_ON
 from homeassistant.core import HomeAssistant, ServiceCall
 
-from tests.components import (
+from tests.components.common import (
     TriggerStateDescription,
     arm_trigger,
     assert_trigger_behavior_any,
