+++ tests/components/vacuum/test_condition.py
@@ -7,7 +7,7 @@ import pytest
 from homeassistant.components.vacuum import VacuumActivity
 from homeassistant.core import HomeAssistant
 
-from tests.components import (
+from tests.components.common import (
     ConditionStateDescription,
     assert_condition_gated_by_labs_flag,
     create_target_condition,
