+++ tests/components/window/test_trigger.py
@@ -8,7 +8,7 @@ from homeassistant.components.cover import ATTR_IS_CLOSED, CoverState
 from homeassistant.const import ATTR_DEVICE_CLASS, CONF_ENTITY_ID, STATE_OFF, STATE_ON
 from homeassistant.core import HomeAssistant, ServiceCall
 
-from tests.components import (
+from tests.components.common import (
     TriggerStateDescription,
     arm_trigger,
     assert_trigger_behavior_any,
